requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Kubernetes客户端
kubernetes==28.1.0

# 配置管理
pyyaml==6.0.1

# 数据验证
pydantic==2.5.0

# 异步支持
aiofiles==23.2.1

# HTTP客户端（用于测试）
httpx==0.25.2